Create a utils file
