+++ KMDFLycaniteFileFilter/KMDFLycaniteFileFilter.vcxproj
@@ -188,6 +188,7 @@
     <ClInclude Include="Kashmap.h" />
     <ClInclude Include="IKashmap.h" />
     <ClInclude Include="Permissions.h" />
+    <ClInclude Include="Utils.h" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
