+++ KMDFLycaniteFileFilter/KMDFLycaniteFileFilter.vcxproj.filters
@@ -46,5 +46,8 @@
     <ClInclude Include="Permissions.h">
       <Filter>Header Files</Filter>
     </ClInclude>
+    <ClInclude Include="Utils.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
   </ItemGroup>
 </Project>
