+++ KMDFLycaniteFileFilter/Kashmap.h
@@ -7,7 +7,7 @@
 #pragma warning(push, 0)
 #pragma warning(disable : 4668)
 #endif
-#include <ntddk.h>
+#include "Utils.h"
 
 #if (defined(_MSC_VER) && defined(__AVX__)) ||                                 \
     (!defined(_MSC_VER) && defined(__SSE4_2__))
@@ -62,21 +62,6 @@ struct hashmap_s {
 #if defined(__cplusplus)
 extern "C" {
 #endif
-
-    // override calloc for kernel
-    PVOID calloc(SIZE_T elementCount, SIZE_T elementSize);
-    
-    // override free for kernel
-    VOID free(PVOID mem);
-
-    // override memcpy for kernel
-    PVOID Kmemcpy(PVOID destination, CONST PVOID source, SIZE_T size);
-
-    // override memcmp
-    INT32 Kmemcmp(CONST PVOID pointer1, CONST PVOID pointer2, SIZE_T size);
-
-    // override memset
-    PVOID Kmemset(PVOID pointer, INT8 value, SIZE_T count);
 
     /// @brief Create a hashmap.
     /// @param initial_size The initial size of the hashmap. Must be a power of two.
@@ -180,55 +165,6 @@ extern "C" {
 #define HASHMAP_PTR_CAST(type, x) ((type)x)
 #define HASHMAP_NULL 0
 #endif
-
-PVOID malloc(SIZE_T memory_size) {
-    return ExAllocatePoolWithTag(NonPagedPoolNx, memory_size, 'aloc');
-}
-
-PVOID calloc(SIZE_T elementCount, SIZE_T elementSize) {
-    SIZE_T total_size = elementCount * elementSize;
-    PCHAR mem = (PCHAR)ExAllocatePoolWithTag(NonPagedPoolNx, elementCount * elementSize, 'aloc');
-
-    // init data allocated with 0 like calloc
-    for (SIZE_T i = 0; mem != NULL && i < total_size; i++) {
-        mem[i] = 0;
-    }
-
-    return (PVOID)mem;
-}
-
-VOID free(PVOID mem) {
-    if (mem != NULL) {
-        ExFreePoolWithTag(mem, 'aloc');
-    }
-}
-
-PVOID Kmemcpy(PVOID destination, CONST PVOID source, SIZE_T size) {
-    for (SIZE_T i = 0; destination != NULL && i < size; i++) {
-        ((PCHAR)destination)[i] = ((PCHAR)source)[i];
-    }
-    
-    return destination;
-}
-
-// override memcmp
-INT32 Kmemcmp(CONST PVOID pointer1, CONST PVOID pointer2, SIZE_T size) {
-    if (pointer1 == pointer2) return 0;
-    for (SIZE_T i = 0; i < size; i++) {
-        if (((PCHAR)pointer1)[i] != ((PCHAR)pointer2)[i]) {
-            return (((PCHAR)pointer1)[i] - ((PCHAR)pointer2)[i]);
-        }
-    }
-    return 0;
-}
-
-// override memset
-PVOID Kmemset(PVOID pointer, INT8 value, SIZE_T count) {
-    for (SIZE_T i = 0; pointer != NULL && i < count; i++) {
-        ((PCHAR)pointer)[i] = value;
-    }
-    return (pointer);
-}
 
 INT8 hashmap_create(CONST UINT32 initial_size,
     struct hashmap_s* CONST out_hashmap) {
