+++ KMDFLycaniteFileFilter/Permissions.h
@@ -6,8 +6,6 @@
 extern "C" {
 #endif
 
-	static UINT32 my_strlen(CONST PCHAR str);
-
 	static UINT64 getFilePermission(CONST PCHAR path, CONST struct hashmap_s* map);
 	static UINT64 getFilePermissionWithFree(CONST PCHAR path, CONST struct hashmap_s* map);
 
@@ -16,16 +14,6 @@ extern "C" {
 #if defined(__cplusplus)
 }
 #endif
-
-UINT32 my_strlen(CONST PCHAR str)
-{
-	UINT32 len = 0;
-
-	for (SIZE_T i = 0; str[i] != '\0'; i++) {
-		len += 1;
-	}
-	return len;
-}
 
 UINT64 getFilePermission(CONST PCHAR path, CONST struct hashmap_s* map)
 {
@@ -36,7 +24,6 @@ UINT64 getFilePermission(CONST PCHAR path, CONST struct hashmap_s* map)
 	}
 	else {
 		PCHAR parent = NULL;
-		/* ajouter un bool ou une connerie du genre pour free parent quand la récursive se fait */
 		if ((parent = hasParentFolder(path, map)) != NULL) {
 			return getFilePermissionWithFree(parent, map);
 		}
@@ -56,11 +43,11 @@ UINT64 getFilePermissionWithFree(CONST PCHAR path, CONST struct hashmap_s* map)
 	}
 	else {
 		PCHAR parent = NULL;
-		/* ajouter un bool ou une connerie du genre pour free parent quand la récursive se fait */
 		if ((parent = hasParentFolder(path, map)) != NULL) {
 			return getFilePermissionWithFree(parent, map);
 		}
 		else {
+			free((PVOID)path);
 			return 0; // return no permissions
 		}
 	}
